requirements.txt
Flask
python-dotenv
groq
PyPDF2
pymongo
Werkzeug
typing
logging
